HHH-16423 LazyInitializationException when calling em.refresh() on entity with lazy and eager children
